+++ src/Inputs/Input.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import FormControl from '@material-ui/core/FormControl/FormControl';
-import FormHelperText from '@material-ui/core/FormHelperText/FormHelperText';
 import PropTypes from 'prop-types';
 import withStyles from '@material-ui/core/styles/withStyles';
 import TextField from '@material-ui/core/TextField/TextField';
@@ -105,12 +103,15 @@ export const TextFieldComponent = ({
   success,
   disabled,
   mask,
+  shrinkLabel,
+  placeholder,
   ...input
 }) => (
   <TextField
     disabled={disabled}
     id="outlined-name"
     name={name}
+    placeholder={placeholder || label}
     label={label}
     onChange={onChange}
     InputLabelProps={{
@@ -118,6 +119,7 @@ export const TextFieldComponent = ({
         root: classes[labelClass({ warning, error, success })],
         focused: classes.cssFocused,
       },
+      shrink: shrinkLabel,
     }}
     InputProps={{
       inputComponent: mask ? TextMaskCustom : undefined,
@@ -153,25 +155,26 @@ const InputUI = withStyles(inputStyles)(
     error,
     success,
     disabled,
+    placeholder,
     mask,
     ...input
   }) => (
-    <FormControl key={key} error={!!error} fullWidth={fullWidth}>
-      <TextFieldComponent
-        classes={classes}
-        label={label}
-        onChange={onChange}
-        value={value}
-        name={name}
-        warning={warning}
-        error={error}
-        success={success}
-        disabled={disabled}
-        mask={mask}
-        {...input}
-      />
-      {error && <FormHelperText>{error}</FormHelperText>}
-    </FormControl>
+    <TextFieldComponent
+      key={key}
+      error={!!error}
+      fullWidth={fullWidth}
+      classes={classes}
+      label={label}
+      placeholder={placeholder}
+      onChange={onChange}
+      value={value}
+      name={name}
+      warning={warning}
+      success={success}
+      disabled={disabled}
+      mask={mask}
+      {...input}
+    />
   )
 );
 export const Input = props => <InputUI {...props} />;
@@ -185,6 +188,8 @@ Input.defaultProps = {
   onChange: null,
   error: false,
   mask: null,
+  shrinkLabel: undefined,
+  placeholder: null,
 };
 
 const maskProp = PropTypes.oneOfType([
@@ -220,6 +225,8 @@ Input.propTypes = {
   name: PropTypes.string.isRequired,
   onChange: PropTypes.func,
   mask: maskProp,
+  shrinkLabel: PropTypes.bool,
+  placeholder: PropTypes.string,
 };
 
 export default Input;