+++ src/Select/Controlnput.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {makeStyles,} from '@material-ui/styles';
+import { makeStyles } from '@material-ui/styles';
 import classNames from 'classnames';
 import { TextFieldComponent } from '../Inputs/Input';
 
@@ -9,7 +9,7 @@ const useStyles = makeStyles({
   },
 });
 
-export function inputComponentReact({ inputRef, className, ...props }) {
+export function InputComponentReact({ inputRef, className, ...props }) {
   const classes = useStyles();
   const newClass = classNames(classes.flex, className);
   return <div ref={inputRef} className={newClass} {...props} />;
@@ -24,9 +24,15 @@ export const Controlnput = ({
   <TextFieldComponent
     fullWidth
     InputProps={{
-      inputComponent: inputComponentReact,
+      inputComponent: InputComponentReact,
       inputProps: {
-        className: selectProps.childrenClasses.input,
+        className: classNames(selectProps.childrenClasses.input, {
+          [selectProps.childrenClasses.inputMultilineDense]:
+            selectProps.isMulti && selectProps.value.length > 0,
+          [selectProps.childrenClasses.inputNormal]:
+            !selectProps.isMulti ||
+            (selectProps.isMulti && selectProps.value.length === 0),
+        }),
         inputRef: innerRef,
         children,
         ...innerProps,